Se creo el diseño del fragment detail

+++ app/src/main/java/com/jonathan/empowerment/data/datasource/local/entity/MetricEntity.kt
@@ -1,10 +1,9 @@
 package com.jonathan.empowerment.data.datasource.local.entity
 
 import com.jonathan.empowerment.domain.model.Metric
-import com.jonathan.empowerment.domain.model.Us
 
 data class MetricEntity(
-    val amount: Long?,
+    val amount: Float?,
     val unitShort: String?,
     val unitLong: String?
 )
