+++ app/src/main/java/com/jonathan/empowerment/data/datasource/local/entity/UsEntity.kt
@@ -3,7 +3,7 @@ package com.jonathan.empowerment.data.datasource.local.entity
 import com.jonathan.empowerment.domain.model.Us
 
 data class UsEntity(
-    val amount: Long?,
+    val amount: Float?,
     val unitShort: String?,
     val unitLong: String?
 )
