+++ app/src/main/java/com/jonathan/empowerment/data/datasource/remote/model/MetricModel.kt
@@ -2,10 +2,9 @@ package com.jonathan.empowerment.data.datasource.remote.model
 
 import com.google.gson.annotations.SerializedName
 import com.jonathan.empowerment.domain.model.Metric
-import com.jonathan.empowerment.domain.model.Us
 
 data class MetricModel(
-    @SerializedName("amount") val amount: Long?,
+    @SerializedName("amount") val amount: Float?,
     @SerializedName("unitShort") val unitShort: String?,
     @SerializedName("unitLong") val unitLong: String?
 )
