+++ app/src/main/java/com/jonathan/empowerment/data/datasource/remote/model/UsModel.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 import com.jonathan.empowerment.domain.model.Us
 
 data class UsModel(
-    @SerializedName("amount") val amount: Long?,
+    @SerializedName("amount") val amount: Float?,
     @SerializedName("unitShort") val unitShort: String?,
     @SerializedName("unitLong") val unitLong: String?
 )
